backend/app/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
autopep8==2.3.1
certifi==2024.7.4
charset-normalizer==3.3.2
fastapi[standard]==0.112.1
idna==3.7
pycodestyle==2.12.1
pydantic==2.8.2
pydantic_core==2.20.1
requests==2.32.3
sniffio==1.3.1
starlette==0.38.2
typing_extensions==4.12.2
urllib3==2.2.2
python-dotenv==1.0.1
sqlalchemy==2.0.32
databases==0.9.0
asyncpg==0.29.0
psycopg2==2.9.9
jwt==1.3.1
passlib==1.7.4
python-jose[cryptography]==3.3.0
python-dateutil==2.9.0.post0
aioredis==2.0.1